Adds action test for add note
